Split monolithic test file into focused modules
Additionally, _skip_if_no_bcftools was renamed to skip_if_no_bcftools and moved to conftest.py alongside the other shared helpers.
